attachments logic to base api-import task
